Rename svg to svgutil
