--- svg/go.mod
@@ -1,3 +1,0 @@
-module github.com/mathpresso/go-utils/svg
-
-go 1.18
