+++ svgutil/LICENSE
