+++ svgutil/go.mod
@@ -1,0 +1,3 @@
+module github.com/mathpresso/go-utils/svgutil
+
+go 1.18
